fix(portal): address Copilot review on propose_remediation
- Wire migration 007 into MIGRATIONS so apply order matches filename
  order (idempotent CREATE TABLE — safe even if stiglab applied it
  first).
- Switch token columns to BIGINT + bind as i64 so multi-billion-token
  responses can't wrap to negative through `as i32`.
- Replace per-call human stub_reason strings with stable machine
  codes (`no_credential`, `credential_lookup_failed`,
  `credential_key_missing`, `credential_decrypt_failed`,
  `budget_exceeded`, `anthropic_init_failed`,
  `anthropic_call_failed`) so the dashboard can branch on reason and
  internal errors stop leaking via chat. Budget code carries
  `details: {spent_usd, cap_usd}` for the UI.
- Drop the response body from the Anthropic client error message —
  log it server-side instead — so non-2xx surprises don't appear in
  user-visible `stub_reason`.
- Bound `collect_recent_log_tails` at the SQL layer (`ORDER BY seq
  DESC LIMIT 64`) so long sessions don't blow up tool latency.
- Clarify `BudgetStatus::Ok` doc (no remaining/headroom field —
  caller computes from spent/cap).
- Clarify `parse_remediation_response` doc to describe the actual
  parse-or-fallback behavior, not embedded-object extraction.
- Update `remediation_monthly_cap_usd` docstring to match the new
  `stub_reason = "budget_exceeded"` + `details` shape.

+++ crates/onsager-portal/migrations/008_portal_remediation_calls.sql
@@ -19,10 +19,10 @@ CREATE TABLE IF NOT EXISTS portal_remediation_calls (
     user_id                     TEXT        NOT NULL,
     artifact_id                 TEXT        NOT NULL,
     model                       TEXT        NOT NULL,
-    input_tokens                INTEGER     NOT NULL DEFAULT 0,
-    output_tokens               INTEGER     NOT NULL DEFAULT 0,
-    cache_creation_input_tokens INTEGER     NOT NULL DEFAULT 0,
-    cache_read_input_tokens     INTEGER     NOT NULL DEFAULT 0,
+    input_tokens                BIGINT      NOT NULL DEFAULT 0,
+    output_tokens               BIGINT      NOT NULL DEFAULT 0,
+    cache_creation_input_tokens BIGINT      NOT NULL DEFAULT 0,
+    cache_read_input_tokens     BIGINT      NOT NULL DEFAULT 0,
     cost_usd                    DOUBLE PRECISION NOT NULL DEFAULT 0,
     created_at                  TIMESTAMPTZ NOT NULL DEFAULT NOW()
 );
